refactor(training): rename corpus_v2 package to corpus_projection, rubric_v2 to rubric
