+++ daydream/training/corpus_projection/__init__.py
@@ -1,0 +1,5 @@
+"""Corpus projection: per-finding training records projected from curated bundles."""
+
+from daydream.training.corpus_projection.projector import BuildFrozenCorpusConfig, build_frozen_corpus
+
+__all__ = ["BuildFrozenCorpusConfig", "build_frozen_corpus"]
