--- daydream/training/corpus_v2/__init__.py
@@ -1,5 +1,0 @@
-"""Corpus v2: per-finding training records projected from curated bundles."""
-
-from daydream.training.corpus_v2.projector import BuildCorpusV2Config, run_build_corpus_v2
-
-__all__ = ["BuildCorpusV2Config", "run_build_corpus_v2"]
