+++ tests/test_stacks_v2_load.py
@@ -16,7 +16,7 @@ from typing import Any, cast
 
 import pytest
 
-from daydream.training.corpus_v2.splits import assign_split
+from daydream.training.corpus_projection.splits import assign_split
 from daydream.training.stacks_v2 import V2Projection, load_v2_projection
 
 SALT = "issue-1081-salt"
