+++ tests/test_training_adjudication_queue.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pytest
 
 from daydream.training.adjudication.queue import build_queue
-from daydream.training.corpus_v2.identity import record_id
+from daydream.training.corpus_projection.identity import record_id
 
 
 def _session(
@@ -85,7 +85,7 @@ def test_digest_drift_reopens_item_and_missing_digest_fails_closed() -> None:
 
 def test_decisive_adjudication_entry_fails_closed(monkeypatch: pytest.MonkeyPatch) -> None:
     from daydream.training.adjudication import queue as queue_module
-    from daydream.training.corpus_v2.projector import project_findings
+    from daydream.training.corpus_projection.projector import project_findings
 
     session = _session("s1", "fp-a", "ambiguous", "d1")
 
@@ -111,7 +111,7 @@ def test_disposition_sets_are_single_sourced() -> None:
     from daydream.training.adjudication.queue import (
         _NON_DECISIVE_DISPOSITIONS as queue_set,
     )
-    from daydream.training.corpus_v2.tiers import (
+    from daydream.training.corpus_projection.tiers import (
         _NON_DECISIVE_DISPOSITIONS as tiers_set,
     )
